feat: add support for 2d and 3D instancing (#337)
* feat: start adding instancing support

* feat: add instancing example

* feat: add 2D instancing

* feat: add support for deformations in instancing

* warning fixes + add prelude

* feat: simplify instancing API

* chore: switch to github workflows for CI

* chore: cargo fmt

* chore: clippy fixes

* chore: fix typos

* fix wasm build

+++ src/window/gl_canvas.rs
@@ -210,7 +210,7 @@ impl AbstractCanvas for GLCanvas {
     }
 
     fn scale_factor(&self) -> f64 {
-        self.window.window().scale_factor() as f64
+        self.window.window().scale_factor()
     }
 
     fn set_title(&mut self, title: &str) {
